Finished refactoring to class fields to set state
Setting a few newObj variables, helps to set the inital state, and also
be able to add another copy of that newObj, further down in another
function (the addNewWork or addNewEdu handlers)

Now I just need to figure out what i will do for handling input on
the workExps and eduExps arrays. I think i can just fit an if statement
inside of handleInput(), and if the parameter passed aims at workExps
or eduExps arrays, then do a different setState and return to outside
the function... Else, do the setState for 'general.propery' (the
setState that we already have written there)

Also I am now thinking that maybe instead of expecting the parameter to
pass the exact string, maybe there is a better way to do this...
dont know... only thing I can think of is enums... maybe yea idk.
Also, I am thinking of taking away the 'exp' parts of the names, bc I
think that leads to better naming.